Change API of `center_of_gravity`
No longer use keywords and rely on a *mass function* to compute the mass
according to the needs of the caller. The default mass function simply
returns the nonnegative part of the weights times the data. To restrict
computations on a smaller region, the caller can use a view or a boxed
array.