A point is no longer an abstract vector
It turns out that having points abstract vectors was a source of
confusion with some common operations like `map` or broadcasted
operators leading to unexpected results and thus a bad idea. A point is
now more like a n-tuple of coordinates, it is iterable and indexable, it
has a length and an element size that are both "traits" (i.e. inferrable
form the type).